# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Soiree API
  version: 0.1.0
  description: |
    Soiree — Robocon test-run video manager HTTP API.

    Phase 1 の主要リソース（masters / sessions / videos / runs / markers / search）を契約として定義する。
    実装は spec-devflow.md §4〜§9 の順に handler / service / UI を縦断で積む。

    ## 共通方針

    - JSON フィールド名は camelCase
    - ID は UUID 文字列
    - 時刻は ISO 8601 (RFC 3339) タイムスタンプ
    - 動画上の時間は秒単位（int = 整数秒、float = 小数秒）
    - リスト系は `{ data: [...], pagination: {...} }` のエンベロープ
    - ページングは cursor ベース (`?cursor=...&limit=...`)
    - エラーは `ErrorResponse` で統一
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same-origin (production / docker)
security:
  - xUserId: []
...
paths:
...
  /teams:
    get:
      tags: [teams]
      operationId: listTeams
      summary: チーム一覧
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamList"
    post:
      tags: [teams]
      operationId: createTeam
      summary: チーム作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
  /teams/{teamId}:
    parameters:
      - $ref: "#/components/parameters/TeamId"
    get:
      tags: [teams]
      operationId: getTeam
      summary: チーム取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [teams]
      operationId: updateTeam
      summary: チーム更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [teams]
      operationId: deleteTeam
      summary: チーム削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /robots:
    get:
      tags: [robots]
      operationId: listRobots
      summary: ロボット一覧
      parameters:
        - name: tournamentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: teamId
          in: query
          description: チームで絞り込む
          required: false
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RobotList"
    post:
      tags: [robots]
      operationId: createRobot
      summary: ロボット作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRobotRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Robot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /tournaments:
    get:
      tags: [tournaments]
      operationId: listTournaments
      summary: Tournament 一覧
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentList"
    post:
      tags: [tournaments]
      operationId: createTournament
      summary: Tournament 作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "400":
          $ref: "#/components/responses/BadRequest"
  /tournaments/{tournamentId}:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
    get:
      tags: [tournaments]
      operationId: getTournament
      summary: Tournament 取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [tournaments]
      operationId: updateTournament
      summary: Tournament 更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [tournaments]
      operationId: deleteTournament
      summary: Tournament 削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    xUserId:
      type: apiKey
      in: header
      name: X-User-Id
      description: |
        現在ユーザーの UUID。セルフホスト・全員同権限の前提のため
        Phase 1 はクライアントが選択ユーザーを header で送る簡易方式。
        将来正式な認証を入れたら securityScheme を差し替える。
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: 前回レスポンスの `pagination.nextCursor`
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: 1 レスポンスあたり最大件数（1〜200、既定 50）
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TournamentId:
      name: tournamentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: 競合（重複や状態不一致）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 機械可読なエラーコード
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Pagination:
      type: object
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: [string, "null"]
          description: 次ページを取得するための cursor

    # ---------- Enums ----------
...
    Team:
      type: object
      required: [id, name, isOwn, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isOwn:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        isOwn:
          type: boolean
          default: false
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        isOwn:
          type: boolean
    TeamList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Team"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---------- Robot ----------
    Robot:
      type: object
      required: [id, tournamentId, teamId, name, version, primaryImageId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
          description: ロボットは (tournament, team) ごとに別レコードとして所有される
        teamId:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
          description: バージョン文字列。空文字列も可
        primaryImageId:
          type: [string, "null"]
          format: uuid
          description: 一覧サムネに使う画像 ID。未設定なら null
        createdAt:
          type: string
          format: date-time
    CreateRobotRequest:
      type: object
      required: [tournamentId, teamId, name]
      properties:
        tournamentId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        version:
          type: string
          default: ""
...
    RobotList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Robot"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---------- Robot Image ----------
...
    Tournament:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    CreateTournamentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
    UpdateTournamentRequest:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
    TournamentList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tournament"
        pagination:
          $ref: "#/components/schemas/Pagination"
...